+++ TABTestKit/Classes/Protocols/Element.swift
@@ -33,6 +33,9 @@ public protocol Element {
 	/// Represents the frame of the element, within the screen.
 	var frameInScreen: CGRect { get }
 	
+	/// The top coordinate of the element, relative to the screen.
+	var topCoordinate: CGVector { get }
+	
 	/// The underlying XCUIElement that this element represents. You should rarely need to access or override this, but you can if you need to.
 	var underlyingXCUIElement: XCUIElement { get }
 	
@@ -48,9 +51,9 @@ public extension Element {
 	
 	var frameInScreen: CGRect { return underlyingXCUIElement.frame }
 	
-	var underlyingXCUIElement: XCUIElement {
-		return parent.underlyingXCUIElement.descendants(matching: type).matching(type, identifier: id).element(boundBy: index)
-	}
+	var topCoordinate: CGVector { return defaultTopCoordinate }
+	
+	var underlyingXCUIElement: XCUIElement { return defaultUnderlyingXCUIElement }
 	
 }